"Refactor index.js to use userController, remove bcrypt and jwt imports, and add dotenv config; update package-lock.json and package.json to include dotenv dependency"
